Content and images are loading

+++ app/src/main/java/com/repkap11/repweather/api/rest/Clouds.java
@@ -13,6 +13,6 @@ public class Clouds {
      */
     @SerializedName("all")
     @Expose
-    public Integer all;
+    public Float all;
 
 }
